requirements.txt
aiofiles==23.2.1
aiogram
aiogram_tools==1.0.1
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
attrs==23.1.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
easyocr==1.7.1
executing==2.0.1
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
gino==1.0.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
icecream==2.1.3
idna==3.6
imageio==2.33.0
imgkit==1.2.3
iso8601==1.1.0
isodate==0.6.1
Jinja2==3.1.2
lazy_loader==0.3
lxml==5.2.1
magic-filter==1.0.12
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.2
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.2.1
passlib==1.7.4
Pillow==10.1.0
pydentic
yt-dlp
platformdirs==4.2.0
pyclipper==1.3.0.post5
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pypika-tortoise==0.1.6
pyTelegramBotAPI==4.14.0
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
scikit-image==0.22.0
scipy==1.11.4
shapely==2.0.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==1.3.24
sympy==1.12
tifffile==2023.9.26
torch==2.1.1
torchvision==0.16.1
tortoise-orm==0.20.0
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.2.1
wget==3.2
yarl==1.9.3
zeep==4.2.1
